requirements.txt
fastapi==0.114.0
langchain-cli==0.0.30
python-dotenv==1.0.1
langchain-community==0.2.16
langchain==0.2.16
langchain-core==0.2.38
tqdm==4.66.5
langchain-openai==0.1.23
faiss-cpu==1.8.0.post1